requirements.txt
langchain-community  # Requires langchain-core>=0.3.0 :cite[2]
langchain-core      # Matches community package requirements
langchain-huggingface
faiss-cpu
sentence-transformers
rank-bm25
numpy
python-docx
pymupdf
pdfplumber
pytesseract
nltk
scikit-learn
huggingface-hub
pillow